[Chore #166097423] Added users app and reorganized
